% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_neighbor.R
\name{find.neighbor}
\alias{find.neighbor}
\title{Identify all neighbors of input feature}
\usage{
find.neighbor(self, other, type = c("right", "left", "top", "bottom"))
}
\arguments{
\item{self}{A vector of (x,y) coordinates}

\item{other}{A vector of (x,y) coordinates}

\item{type}{Either "right", "left", "top", or "bottom"}
}
\value{
An indicator of whether self and other are neighbors.
}
\description{
Identify all neighbors of input feature
}
\details{
write XXX
}
\examples{
# two dimensional location
find.neighbor(c(1,1),c(1,2), type = "top")
find.neighbor(c(1,1),c(1,2), type = "bottom")

}
\seealso{
\code{\link{region.plot.2D}, \link{build.wall}}
}
\author{
Hung-Ching Chang
}
