% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipid1D.R
\name{lipid1D}
\alias{lipid1D}
\title{1-dimensional TAG lipid data}
\usage{
data(lipid1D)
}
\value{
A data frame with 15 rows (features) and 85 columns:
\itemize{
  \item{totallength: }{total chain length}
  \item{ctrl: }{control group}
  \item{exp: }{experimental group}
  ...
}
}
\description{
A subset of data from XXX
}
\keyword{dataset,}
\keyword{lipid1D}
