% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipid2D.R
\name{lipid2D}
\alias{lipid2D}
\title{2-dimensional TAG lipid data}
\usage{
data(lipid2D)
}
\value{
A data frame with 93 rows (features) and 86 columns:
\itemize{
  \item{totallength: }{total chain length}
  \item{totaldb: }{total double bond}
  \item{ctrl: }{control group}
  \item{exp: }{experimental group}
  ...
}
}
\description{
A subset of data from XXX
}
\keyword{dataset,}
\keyword{lipid2D}
