% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LipidFete_test.R
\name{LipidFete.test}
\alias{LipidFete.test}
\title{Calculate p-value of lipidomic features by permutation test}
\usage{
LipidFete.test(
  X,
  X.info,
  group,
  radius = 3,
  own.contri = 0.5,
  x.distance = 1,
  y.distance = 1,
  dimension = c(1, 2),
  permute.time = 1e+05,
  seed = 1234,
  ...
)
}
\arguments{
\item{X}{A \code{matrix} of lipid expression data. Row is samples, and column is
lipid feature such as double bond and chain length.}

\item{X.info}{A \code{data.frame} or \code{matrix} of lipid feature names. (Only allow 1 or 2 columns)}

\item{group}{A vector of observed grouping variable. (Only allow 0 and 1)}

\item{radius}{A value to decide the number of neighbor included}

\item{own.contri}{A proportion of own contribution (suggest 0.5 ~ 1)}

\item{x.distance}{A value of unit distance in x-axis}

\item{y.distance}{A value of unit distance in y-axis}

\item{dimension}{A value to indicate the dimension of lipid features (1 or 2)}

\item{permute.time}{A value of permutations}

\item{seed}{A value of random seed index}
}
\value{
A \code{data.frame} of containing lipidomic feature testing result.
\itemize{
    \item{X.info: }{lipidomic feature information.}
    \item{direction: }{the feature expression level in group 1 is higher (+) or lower (-).}
    \item{smoothing.pval.BH: }{p-value of lipidomic features by permutation test with BH correction.}
    \item{log2.FC: }{log2-transformed fold change.}
}
}
\description{
Calculate p-value of lipidomic features by permutation test.
The test of region statistics smoothing with Gaussian kernel integrates
neighbor's information and provides a stable testing result under small sample size.
}
\details{
write XXX
}
\examples{
data(lipid2D)
X <- t(as.matrix(lipid2D[,-c(1:2)]))
X.info <- lipid2D[,1:2]
group <- rep(c(0, 1), c(52,32))
LipidFete.test(X = X,
               X.info = X.info,
               group = group,
               radius = 3,
               own.contri = 0.5,
               x.distance = 2,
               y.distance = 1,
               dimension = 2,
               permute.time = 100000)

}
\seealso{
\code{\link{region.stat}, \link{region.plot.1D}, \link{region.plot.2D}}
}
\author{
Hung-Ching Chang
}
