% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pval_annotation.R
\name{pval.annotation}
\alias{pval.annotation}
\title{Annotate p-value with four thresholds}
\usage{
pval.annotation(pval)
}
\arguments{
\item{pval}{A vector of p-value}
}
\value{
An \code{character} of p-value annotation:
\itemize{
    \item{*** }{represents p-value < 0.001.}
    \item{** }{represents 0.001 < p-value < 0.01.}
    \item{* }{represents 0.01 < p-value < 0.05.}
    \item{∆ }{represents 0.05 < p-value < 0.1.}
}
}
\description{
Annotate p-value with four thresholds.
}
\details{
write XXX
}
\examples{
pval <- 0.04
pval.annotation(pval)

}
\seealso{
\code{\link{region.plot.2D}}
}
\author{
Hung-Ching Chang
}
