% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_plot_1D.R
\name{region.plot.1D}
\alias{region.plot.1D}
\title{Visualize LipidFete test result of 1D feature}
\usage{
region.plot.1D(
  X,
  X.info,
  group,
  direction,
  smoothing.pval,
  marginal.pval,
  feature.name,
  cut.point = 0.05,
  ...
)
}
\arguments{
\item{X}{A \code{matrix} of lipid expression data. Row is samples, and column is
lipid feature such as double bond and chain length.}

\item{X.info}{A \code{data.frame} or \code{matrix} of lipid feature names. (Only allow 1 or 2 columns)}

\item{group}{A vector of observed grouping variable. (Only allow 0 and 1)}

\item{direction}{A vector of direction of feature expression level}

\item{smoothing.pval}{A vector of kernel smoothing p-value}

\item{marginal.pval}{A vector of marginal p-value}

\item{feature.name}{A vector of log2-transformed fold change}

\item{cut.point}{A cut point for smoothing.pval}
}
\value{
A figure presenting significant lipidomic features colored by direction of feature expression level
}
\description{
Visualize LipidFete test result of 1D feature for identifying tendencies of lipid features.
}
\details{
write XXX
}
\examples{
data(lipid1D)
X <- t(as.matrix(lipid1D[,2:85]))
X.info <- lipid1D[,1]
group <- rep(c(0, 1), c(52,32))
test.result <- LipidFete.test(X = X,
                              X.info = X.info,
                              group = group,
                              radius = 2,
                              own.contri = 0.5,
                              x.distance = 2,
                              y.distance = 1,
                              dimension = 1,
                              permute.time = 100000)

region.plot.1D(X = X,
               X.info = X.info,
               group = group,
               direction = test.result$direction,
               smoothing.pval = test.result$smoothing.pval.BH,
               marginal.pval = test.result$marginal.pval.BH,
               feature.name = colnames(X)[1],
               cut.point = 0.05)

}
\seealso{
\code{\link{region.stat}, \link{LipidFete.test}, \link{region.plot.2D}}
}
\author{
Hung-Ching Chang
}
