% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_plot_2D.R
\name{region.plot.2D}
\alias{region.plot.2D}
\title{Visualize LipidFete test result of 2D feature}
\usage{
region.plot.2D(
  X.info,
  direction,
  smoothing.pval,
  marginal.pval,
  log2.FC,
  cut.point = 0.05,
  x.distance = 1,
  y.distance = 1,
  log2.FC.cut.point = 3,
  ...
)
}
\arguments{
\item{X.info}{A \code{data.frame} or \code{matrix} of lipid feature names. (Only allow 1 or 2 columns)}

\item{direction}{A vector of direction of feature expression level}

\item{smoothing.pval}{A vector of kernel smoothing p-value}

\item{marginal.pval}{A vector of marginal p-value}

\item{log2.FC}{A vector of log2-transformed fold change}

\item{cut.point}{A cut point for smoothing.pval}

\item{x.distance}{A value of unit distance in x-axis}

\item{y.distance}{A value of unit distance in y-axis}

\item{log2.FC.cut.point}{A cut point for log2-transformed fold change}
}
\value{
A heatmap presenting significant lipidomic features, colored by log2 fold change, and annotated by marginal p-value.
}
\description{
Visualize LipidFete test result of 2D feature for identifying tendencies of lipid features.
}
\details{
write XXX
}
\examples{
data(lipid2D)
X <- t(as.matrix(lipid2D[,-c(1:2)]))
X.info <- lipid2D[,1:2]
group <- rep(c(0, 1), c(52,32))
test.result <- LipidFete.test(X = X,
                              X.info = X.info,
                              group = group,
                              radius = 3,
                              own.contri = 0.5,
                              x.distance = 2,
                              y.distance = 1,
                              dimension = 2,
                              permute.time = 100000)

region.plot.2D(X.info = X.info,
               direction = test.result$direction,
               smoothing.pval = test.result$smoothing.pval.BH,
               marginal.pval = test.result$marginal.pval.BH,
               log2.FC = test.result$log2.FC,
               cut.point = 0.05,
               x.distance = 2,
               y.distance = 1)

}
\seealso{
\code{\link{LipidFete.test}, \link{build.wall}, \link{pval.annotation}}
}
\author{
Hung-Ching Chang
}
