% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_stat.R
\name{region.stat}
\alias{region.stat}
\title{Calculate region statistics}
\usage{
region.stat(X, Y, ...)
}
\arguments{
\item{X}{A \code{matrix} of lipid expression data. Row is samples, and column is
lipid feature such as double bond and chain length.}

\item{Y}{A \code{matrix} of observed or permuted grouping variable. Each column is
a grouping result. Only allow 0 and 1.}
}
\value{
A \code{matrix} of region statistics calculated as described above.
Each column is a set of region statistics corresponding with each grouping result.
}
\description{
Calculate region statistics of every lipidomic feature. The region statistics
is signed p-value, which is constituted by p-value of marginal test and sign of mean difference
between groups.
}
\details{
write XXX
}
\examples{
# one dimensional Y.
X0 <- matrix(rnorm(100,0,1), nrow = 10)
X1 <- matrix(rnorm(100,2,1), nrow = 10)
X <- rbind(X0,X1)
Y <- matrix(rep(c(0,1), each = 10), ncol = 1)

# calculate region statistics
rstat <- region.stat(X = X, Y = Y)

# high dimensional Y.
X0 <- matrix(rnorm(100,0,1), nrow = 10)
X1 <- matrix(rnorm(100,2,1), nrow = 10)
X <- rbind(X0,X1)
group <- rep(c(0,1), each = 10)
Y <- sapply(seq_len(20),
            function(i) sample(group, replace = F))

# calculate region statistics
rstat <- region.stat(X = X, Y = Y)

}
\seealso{
\code{\link{LipidFete.test}}
}
\author{
Hung-Ching Chang
}
